some rearraging of the code parts
